changed population to use type double instead of long or ulong.  All the calculations we being done using double and just converted to long for storing/displaying.  Added unnecessary casting so removed it all

+++ GurpsSpace/PlanetCreation/IPlanetCreator.cs
@@ -24,7 +24,7 @@ namespace GurpsSpace.PlanetCreation
 
         int SetLocalTechLevel(ViewModelPlanet p);
 
-        ulong SetPopulation(ViewModelPlanet p);
+        double SetPopulation(ViewModelPlanet p);
 
         (eWorldUnityLevel, fGovernmentSpecialConditions) SetWorldGovernance(ViewModelPlanet p);
